Fix Ical bug.
