+++ app/Contracts/FacultyContract.php
@@ -5,8 +5,12 @@ interface FacultyContract{
     // public function editPost($postData, $id);
 
     public function getClassList($term,$email);
+
     public function getFinalExamTimes($term,$email);
+
     public function getOfficeHours($term,$email);
+    
+    public function getIcal($instructorInfo);
 
 }
 ?>
